data-simulation/meeting-request-generator/requirements.txt
fastapi
uvicorn
faker
kafka-python
apscheduler